connected the user table with task table OneToMany

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './tasks.repository';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,24 +15,33 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) {}
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const taskFound = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User
+    ): Promise<Task> {
+        const taskFound = await this.taskRepository.findOne({ where: {id, userId: user.id }});
         if (!taskFound) {
             throw new NotFoundException('Task with ID '+ id + " not found");
         } 
         return taskFound;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTaskById(id: number): Promise<number> {
-        const result = await this.taskRepository.delete(id)
+    async deleteTaskById(
+        id: number, 
+        user: User
+    ): Promise<number> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
         
         if (result.affected === 0) {
             throw new NotFoundException('Task with ID '+ id + " not found");
@@ -39,8 +49,12 @@ export class TasksService {
         return result.affected;
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User
+    ): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await this.taskRepository.save(task);
         return task;
